requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
qrcode>=7.4.0
wxPython>=4.2.0
PyYAML>=6.0
lxml